refactor FeatureTrack
Make it keep its IntervalStore object encapsulated, and delegate
methods to it as appropriate.  Also, give it a DESTROY that calls
finishLoad() if a load is in progress, obviating the need to
explicitly call finishLoad() most of the time.  Also, make the
encapsulated NameHandler not be created until startLoad() is called.